requirements.txt
django
djangorestframework
gunicorn
psycopg2-binary
django-environ
django-cors-headers
httpx
mcp
croniter